refactor(coyote): overall cleansing
