+++ tomcat/src/main/java/org/apache/coyote/response/ResourceView.java
@@ -4,10 +4,9 @@ import java.io.File;
 import java.io.IOException;
 import java.nio.file.Files;
 import java.nio.file.Path;
-import org.apache.coyote.exception.ExceptionPage;
 import org.apache.coyote.response.HttpResponse.HttpResponseBuilder;
 import org.apache.coyote.support.HttpStatus;
-import org.apache.coyote.exception.HttpException;
+import org.apache.coyote.support.HttpException;
 
 public class ResourceView {
