+++ tomcat/src/main/java/org/apache/coyote/servlet/CustomServlet.java
@@ -1,8 +1,7 @@
 package org.apache.coyote.servlet;
 
-import org.apache.coyote.exception.HttpException;
+import org.apache.coyote.support.HttpException;
 import org.apache.coyote.request.HttpRequest;
-import org.apache.coyote.request.HttpRequestHandler;
 import org.apache.coyote.response.HttpResponse;
 
 public class CustomServlet {
