+++ tomcat/src/main/java/org/apache/coyote/support/HttpMethod.java
@@ -1,7 +1,5 @@
 package org.apache.coyote.support;
 
-import org.apache.coyote.exception.HttpException;
-
 public enum HttpMethod {
     GET, POST, PUT, DELETE;
